requirements.txt
Django==1.8.3
Pillow==2.9.0
Unidecode==0.04.18
Willow==0.2.1
argparse==1.2.1
beautifulsoup4==4.4.0
django-appconf==1.0.1
django-bower==5.0.4
django-compressor==1.5
django-disqus==0.5
django-filer==0.9.12
django-model-utils==2.0.3
django-modelcluster==0.6.2
django-mptt==0.7.4
django-polymorphic==0.7.1
django-taggit==0.16.2
django-treebeard==3.0
djangorestframework==3.1.3
easy-thumbnails==2.2
elasticsearch==1.6.0
funcsigs==0.4
future==0.15.0
html5lib==0.999
jsonfield==1.0.3
mock==1.3.0
numpy==1.9.2
pandas==0.16.2
pbr==1.3.0
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
rest-pandas==0.3.2
six==1.9.0
urllib3==1.11
wagtail==1.0
wsgiref==0.1.2
